requirements.txt
streamlit
requests
openai
python-dotenv
langchain
langchain-openai
langchain-core
